new tap values
we changed the tap values to 24, 32, 45, 60, 125, 200 while changing the mux frequency to 800 and the time stamp to output correctly (It was pulling the period variable from somewhere besides sensors.c).  We also included the stutter code as well as changed the impedance equation to match our new method with using adc value instead of gain.  This is set up for the callibration of board #4 currently.

+++ Sensors/Serializer.c
@@ -14,7 +14,7 @@ int serializer_isFull() {
     return index == 0;
 }
 
-void serializer_setTimestamp(uint16_t timestamp) {
+void serializer_setTimestamp(uint32_t timestamp) {
     sensorData.timestamp = timestamp;
 }
 
@@ -24,9 +24,9 @@ void serializer_addImpedance(float impedance) {
 }
 
 int serializer_serialize(char* buffer) {
-    memcpy(buffer, &sensorData.timestamp, sizeof(unsigned short));
-    memcpy(buffer + sizeof(unsigned short), sensorData.impedanceValues, sizeof(float) * NUM_SENSORS);
-    return sizeof(unsigned short) + sizeof(float) * NUM_SENSORS;
+    memcpy(buffer, &sensorData.timestamp, sizeof(uint32_t));
+    memcpy(buffer + sizeof(uint32_t), sensorData.impedanceValues, sizeof(float) * NUM_SENSORS);
+    return sizeof(uint32_t) + sizeof(float) * NUM_SENSORS;
 }
 
 int serializer_serializeReadable(char* buffer) {
